+++ packages/picasso.js/src/core/component/component-factory.js
@@ -53,9 +53,6 @@ function prepareContext(ctx, definition, opts) {
   }
 
   // TODO add setters and log warnings / errors to console
-  Object.defineProperty(ctx, 'userSettings', {
-    get: settings
-  });
   Object.defineProperty(ctx, 'settings', {
     get: settings
   });
@@ -240,17 +237,15 @@ function componentFactory(definition, context = {}) {
   const instanceContext = extend({}, config);
 
   // Create a callback that calls lifecycle functions in the definition and config (if they exist).
-  function createCallback(method, defaultMethod = () => { }, canBeValue = false, alternateMethod) {
+  function createCallback(method, defaultMethod = () => { }, canBeValue = false) {
     return function cb(...args) {
-      const inDefinition = typeof definition[method] !== 'undefined' || typeof definition[alternateMethod] !== 'undefined';
-      const inConfig = typeof config[method] !== 'undefined' || typeof config[alternateMethod] !== 'undefined';
+      const inDefinition = typeof definition[method] !== 'undefined';
+      const inConfig = typeof config[method] !== 'undefined';
 
       let returnValue;
       if (inDefinition) {
         if (typeof definition[method] === 'function') {
           returnValue = definition[method].call(definitionContext, ...args);
-        } else if (typeof definition[alternateMethod] === 'function') {
-          returnValue = definition[alternateMethod].call(definitionContext, ...args);
         } else if (canBeValue) {
           returnValue = definition[method];
         }
@@ -259,8 +254,6 @@ function componentFactory(definition, context = {}) {
       if (inConfig) {
         if (typeof config[method] === 'function') {
           returnValue = config[method].call(instanceContext, ...args);
-        } else if (typeof config[alternateMethod] === 'function') {
-          returnValue = config[alternateMethod].call(instanceContext, ...args);
         } else if (canBeValue) {
           returnValue = config[method];
         }
@@ -274,7 +267,7 @@ function componentFactory(definition, context = {}) {
     };
   }
 
-  const preferredSize = createCallback('preferredSize', () => 0, true, 'getPreferredSize');
+  const preferredSize = createCallback('preferredSize', () => 0, true);
   const resize = createCallback('resize', ({ inner }) => inner);
   const created = createCallback('created');
   const beforeMount = createCallback('beforeMount');
@@ -372,12 +365,12 @@ function componentFactory(definition, context = {}) {
     }
     instanceContext.rect = extend(true, {
       computedPhysical: size.computedPhysical,
-      computedOuter: outer.computed,
-      computedInner: inner.computed
+      computedOuter: outer.computed || outer,
+      computedInner: inner.computed || inner
     }, inner);
     size = extend(true, {
-      computedOuter: outer.computed,
-      computedInner: inner.computed
+      computedOuter: outer.computed || outer,
+      computedInner: inner.computed || inner
     }, size);
   };