+++ packages/picasso.js/src/web/text-manipulation/line-break-resolver.js
@@ -1,7 +1,8 @@
 import extend from 'extend';
 import { breakAll, breakWord } from './word-break';
-import { DEFAULT_LINE_HEIGHT, ELLIPSIS_CHAR } from './text-const';
+import { ELLIPSIS_CHAR } from './text-const';
 import { includesLineBreak } from './string-tokenizer';
+import { fontSizeToLineHeight } from './font-size-to-height';
 
 function generateLineNodes(result, item, halfLead, height) {
   const container = { type: 'container', children: [] };
@@ -71,16 +72,13 @@ export function onLineBreak(measureText) {
       }
 
       const tm = measureText(item);
-      if (tm.width <= item.maxWidth && !includesLineBreak(item.text)) {
-        return;
+      if (tm.width > item.maxWidth || includesLineBreak(item.text)) {
+        const diff = fontSizeToLineHeight(item) - tm.height;
+        const halfLead = diff / 2;
+        const result = wordBreakFn(item, wrappedMeasureText(item, measureText));
+
+        state.node = generateLineNodes(result, item, halfLead, tm.height); // Convert node to container
       }
-
-      const lineHeight = tm.height * Math.max((isNaN(item.lineHeight) ? DEFAULT_LINE_HEIGHT : item.lineHeight), 0);
-      const diff = lineHeight - tm.height;
-      const halfLead = diff / 2;
-      const result = wordBreakFn(item, wrappedMeasureText(item, measureText));
-
-      state.node = generateLineNodes(result, item, halfLead, tm.height); // Convert node to container
     }
   };
 }