+++ packages/picasso.js/src/web/text-manipulation/text-const.js
@@ -1,9 +1,7 @@
-const DEFAULT_LINE_HEIGHT = 1.2;
 const HYPHENS_CHAR = '\u2010';
 const ELLIPSIS_CHAR = '…';
 
 export {
-  DEFAULT_LINE_HEIGHT,
   HYPHENS_CHAR,
   ELLIPSIS_CHAR
 };