+++ plugins/hammer/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.25.1](https://github.com/qlik-oss/picasso.js/compare/v0.25.0...v0.25.1) (2019-04-29)
+
+**Note:** Version bump only for package picasso-plugin-hammer
+
+
+
+
+
+# [0.25.0](https://github.com/qlik-oss/picasso.js/compare/v0.24.0...v0.25.0) (2019-04-26)
+
+**Note:** Version bump only for package picasso-plugin-hammer
+
+
+
+
+
+# [0.24.0](https://github.com/qlik-oss/picasso.js/compare/v0.23.2...v0.24.0) (2019-04-17)
+
+**Note:** Version bump only for package picasso-plugin-hammer
+
+
+
+
+
+## [0.23.2](https://github.com/qlik-oss/picasso.js/compare/v0.23.1...v0.23.2) (2019-04-08)
+
+**Note:** Version bump only for package picasso-plugin-hammer
+
+
+
+
+
+## [0.23.1](https://github.com/qlik-oss/picasso.js/compare/v0.23.0...v0.23.1) (2019-04-08)
+
+**Note:** Version bump only for package picasso-plugin-hammer
+
+
+
+
+
+# [0.23.0](https://github.com/qlik-oss/picasso.js/compare/v0.22.1...v0.23.0) (2019-04-05)
+
+**Note:** Version bump only for package picasso-plugin-hammer
+
+
+
+
+
 ## [0.22.1](https://github.com/qlik-oss/picasso.js/compare/v0.22.0...v0.22.1) (2019-04-04)
 
 **Note:** Version bump only for package picasso-plugin-hammer